Add files via upload
In the first part of the project I've used three search heuristic to find an approximate solution of the Traveling Salesman Problem (TSP), which is NP-hard. In TSP, we are given a non-negative zero-diagonal symmetric matrix which models a fully connected graph - the nodes are locations, and edge weight between the two nodes is a cost of traveling from one node to the other. Our goal with TSP is to find a way for our Salesman to visit each node once (and only once) before returning to the starting node, while minimizing the cost of doing so. The search heuristics used are Nearest Neighbor Algorithm (NN), NN with 2-opt, Repeated Randomized NN (RRNN), RRNN with 2-opt and NN with 3-opt.

In the second part I have approximated TSP solution using a Minimum Spanning Tree (MST) and A* heuristics. For computing the MST I've used Prim's Algorithm.

In the third part I've used three local search algorithms to find a good approximation of TSP problem. The algorithms used are Hill-Climbing Algorithm, Simulated Annealing and a Genetic Algorithm.

Throughout the project I've performed a detailed analysis and comparison of the mentioned algorithms, visualizing the results.